Added new autoencoder model and added section for taking a single slice from each participant for training the autoencoder.

+++ prepare_data.py
@@ -3,6 +3,7 @@ import numpy as np
 import nibabel as nib
 from create_data_df import create_data_df
 import os
+import h5py
 from helpers import nii2Numpy, normalize
 
 
@@ -67,33 +68,102 @@ def count_valid_samples(imagesDf):
 
 
 
-def prepare_data(mainDir, data_dir, report_dir, input_str, ext, labelName, idColumn, covPath, ratios=[0.75,0.15], batch_size=2, feature_dim=2, pad_base=0):
-    # data = {}
-    # # Setup file names and imagesDfput directories
-    # data["train_x_fn"] = os.path.join(data_dir,'train_x')
-    #
-    # data["train_y_fn"] = os.path.join(data_dir,'train_y')
-    # data["validate_x_fn"] = os.path.join(data_dir,'validate_x')
-    #
-    # data["validate_y_fn"] = os.path.join(data_dir,'validate_y')
-    # data["test_x_fn"] = os.path.join(data_dir,'test_x')
-    #
-    # data["test_y_fn"] = os.path.join(data_dir,'test_y')
+def prepare_data(mainDir, data_dir, report_dir, input_str, ext, labelName, idColumn, covPath, imagesDfOut,ratios=[0.75,0.15]):
 
-    #if images_fn == None: images_fn = os.path.join(report_dir,'images.csv')
+    data = {}
 
-    imagesDf = create_data_df(mainDir, input_str, ext, labelName, covPath, idColumn)
+    # Setup file names and imagesDfput directories
+    data["x_train_fn"] = os.path.join(mainDir,'x_train')
+    data["y_train_fn"] = os.path.join(mainDir,'y_train')
+    data["x_validate_fn"] = os.path.join(mainDir,'x_validate')
+    data["y_validate_fn"] = os.path.join(mainDir,'y_validate')
+    data["x_test_fn"] = os.path.join(mainDir,'x_test')
+    data["y_test_fn"] = os.path.join(mainDir,'y_test')
+    data["image_dim"] = [256,256]
+
+    if os.path.exists(imagesDfOut):
+        imagesDf = pd.read_csv(imagesDfOut)
+        return [imagesDf, data]
+
+
+    # Create the path, label, category data frame
+    imagesDf = create_data_df(mainDir, data_dir, input_str, ext, labelName, covPath, idColumn)
 
     # fix to make sure enough test/train/validate
     attribute_category(imagesDf, 'train', 'labels', ratios[0])
     attribute_category(imagesDf, 'validate','labels', ratios[1])
     imagesDf.category.loc[ imagesDf.category == "unknown" ] = "test"
 
+    # Create npy arrays with dataset split
+    hdf5_path = os.path.join(mainDir, 'datasetSplit.hdf5')
+    #create_hd5(imagesDf,data, hdf5_path)
 
     fname = os.path.join(mainDir, 'sMRI_{}.csv'.format(labelName))
     imagesDf.to_csv(fname, sep=',')
 
-    return imagesDf
+    return [imagesDf,data]
+
+
+def create_hd5(imagesDf,data,hdf5_path):
+
+
+    if os.path.exists(hdf5_path):
+        os.remove(hdf5_path)
+
+    x_train = imagesDf['paths'][imagesDf.category=='train']
+    y_train = imagesDf['labels'][imagesDf.category=='train']
+    x_val = imagesDf['paths'][imagesDf.category=='validate']
+    y_val = imagesDf['labels'][imagesDf.category=='validate']
+    x_test = imagesDf['paths'][imagesDf.category=='test']
+    y_test = imagesDf['labels'][imagesDf.category == 'test']
+
+
+    train_shape = [x_train.shape[0]*256,256,256,1]
+    val_shape = [x_val.shape[0]*256, 256, 256,1]
+    test_shape = [x_test.shape[0]*256, 256, 256,1]
+
+    # open hdf5 file and create arrays
+    hdf5_f = h5py.File(hdf5_path, mode='w')
+    hdf5_f.create_dataset("train_img", train_shape, dtype='float16')
+    hdf5_f.create_dataset("validate_img", val_shape, dtype='float16')
+    hdf5_f.create_dataset("test_img", test_shape, dtype='float16')
+
+    hdf5_f.create_dataset("train_labels", [y_train.shape[0]*train_shape[1],1], np.int8)
+    hdf5_f.create_dataset("validate_labels", [y_val.shape[0]*val_shape[1],1], np.int8)
+    hdf5_f.create_dataset("test_labels", [y_test.shape[0]*test_shape[1],1], np.int8)
+
+    total_index = {'train': 0,
+                   'test': 0,
+                   'validate':0}
+
+    for index, row in imagesDf.iterrows():
+        print('Index: {}'.format(index))
+        if index % 10 == 0: print("Saving",imagesDf["category"][0],"images:",index, '/', imagesDf.shape[0] , end='\r')
+
+        img3D = nii2Numpy(row.paths)
+        label = row.labels
+        img3D = normalize(img3D)
+        img3D.reshape(list(img3D.shape) + [1])
+        labels = np.full((img3D.shape[1], 1), label)
+
+        for j in range(img3D.shape[1]):
+            if img3D[:,j,:].sum() != 0:
+                hdf5_f[row.category + "_img"][(total_index[row.category])] = img3D[:,j,:]
+                hdf5_f[row.category + "_labels"][(total_index[row.category])] = labels[j]
+                total_index[row.category] += 1
+
+    np.save(data['x_train_fn'] + '.npy', hdf5_f['train_img'])
+    np.save(data['x_validate_fn'] + '.npy',hdf5_f['validate_img'])
+    np.save(data['x_test_fn'] + '.npy',hdf5_f['test_img'])
+
+    np.save(data['y_train_fn'] +'.npy',hdf5_f['train_labels'])
+    np.save(data['y_validate_fn'] +'.npy',hdf5_f['validate_labels'])
+    np.save(data['y_test_fn'] +'.npy',hdf5_f['test_labels'])
+    hdf5_f.close()
+    print("")
+    return 0
+
+
 
 
 def unitTest1():
